BUG: PVStateSignal must be exempt from GroupDevice orpaning REF: Move PVStateSignal to signal module to avoid circular import
